Make the Publishable.per_language_publication member visible and editable.
Untill now, this field was set by model designers, as a default value, and
editors couldn't modify it. The main reason for the change is the somewhat
frequent case of having several variations of an image containing localized
text. Allowing users to toggle this field makes it possible to constrain each
variation to one (or more) languages, and all code that checks for a publication
state (ie. the TextBlock or FileListing models of the blocks extension) will
transparently honor this choice.