View factories now set all received parameters as attributes of the produced view.
This mainly affects handler defined as functions, which so far had total control over these attributes. Now they receive a dictionary, instead of keyword arguments, and any value left in the dictionary will be set as an attribute of the view they return (so they can pop specific keys or event empty the whole dictionary if they want to control this behavior).