Fixed changeset 870e723390c9
The aim of the original changeset was to avoid attempting to remove websites
twice from their designated content when said content was deleted (once by the
'unrelated' event, a second time when item deletion cleared the collection).
Nevertheless, the check to limit the 'unrelated' event to non deleted objects
was done using 'is_inserted' (instead of 'is_deleted'). That glitch disabled the
event handler for objects that had not been inserted yet.