Fix for the click delegation behavior of PublishableListing.
The PublishabeListing view allows concrete implementations to include labels
for each entry that, when clicked, will invoke the same behavior of the entry's
view. To do so, the view artificially triggers a click event on a designated
target. Previously, this was done using jQuery's click() method, but it seems
that method only invokes custom handlers, bypassing the element's default
behavior (so a regular link or button didn't work). To fix this, the call to
that method has been replaced with a direct call to the click() method of the
underlying DOM element, which does activate the element's default behavior.