CMS.translate_uri() didn't translate the descriptive part of URIs.
The translate_uri() method (which is used in the language selector and to
generate <link rel="alternate"> tags, among other places) only changed the URI
component that identified the language, but left the rest of the URI untouched.
Since the descriptive ID publication scheme includes the title of the
publishable element as part of the URI, that should be translated as well.
Otherwise, users would suffer a redirection (the CMS would redirect to the
canonical URI for the translation) and Google expects <link rel="alternate">
with hreflang attributes to be bidirectional and to match exactly (it doesn't
deal with redirections).