staticsite extension: use a single invocation of Wget to fetch all URLs
The Wget snapshoter spawned a separate process for each URL that it was required
to fetch. This had two problems: bad performance, and possibly fetching the same
URL over and over (since each process didn't know the set of URLs obtained by
previous calls).

As a side benefit, the snapshoter now also accepts string based URIs, as well as
references to publishable items.