staticsite extension: make it possible to know if the site is being exported
Controllers and views can now call the generating_static_site() function to
determine if the current HTTP request is being issued by a site snapshoter.
This is accomplished by having snapshoters (meaning the Wget snapshoter, which
is the only implementation right now) pass in a header
(X-Woost-StaticSiteSnapshoter) in all their requests.