Relocated CMSController.language and CMSController.authentication.
The language and authentication policies for a Woost application were, until
now, confined to the CMSController class. This made it difficult to access them
throughout the codebase (ie. from a script or an interactive console). A notable
example of this problem was that Publishable.get_uri() didn't rely on the
LanguageModule class to translate URIs into specifc languages, but hardcoded the
logic itself instead. To solve these shortcommings, both policy classes have
been relocated as properties of the woost.app object.

Also, the woost.controllers.module.Module class, which served as a base for the
LanguageModule and AuthenticationModule classes, was a useless historical
artifcat and has been removed.

The old APIs and modules should still work, but they have been deprecated and
will issue the appropiate warnings.