Removed the workflow extension.
The design of the workflow extension doesn't mix well with the blocks extension.
Since the blocks extension will be merged into the core, this extension will
have to be removed.

In the future, a better approach to workflow would be to integrate it with a
"working copies" system, allowing sets of disjoint items (ie. revisions) to
participate together in state transitions.