Removed Item.owner to make per-role caching possible.
The Item.owner member was rarely used; even sites that required the concept of
object ownership ended up implementing their own model, based on the
requirements of their application's domain (for example, "editors can only edit
schools belonging to the districts assigned to their roles).

At the same time, the possibility of basing permissions on object ownership
meant that users with the same set of roles could have different permissions
on the same object. This affected caching very negatively, since per user
caching is clearly not desirable: the hit rate of the resulting cache would be
very low, while memory consumption would rise quickly.

Per role caching, instead, is a much more attractive prospect, but it is
incompatible with the concept of object ownership. By removing this concept,
permissions can be determined solely based on roles, and this makes it possible
to cache content on a per role basis (by adding the roles for the active user to
the caching key of each piece of content in the cache).

Therefore, this changeset removes the Item.owner member, as well as the
OwnershipExpression and OwnItemsFilter classes.