Upgraded TinyMCE to the 4.x version.
As part of the upgrade, the TinyMCE distribution now should be placed in the
resources/scripts folder of the Cocktail package, not in Woost (it was placed
in Woost for historical reasons, but obviously that made no sense).

Among other things, the upgrade makes it possible to copy HTML between editor
instances (the code will be sanitized). On the other hand, the ability to
synchronize the resizing of editors for different translations has been lost
as a result of the upgrade (the previous implementation was an extremely fragile
piece of monkey patching abuse), so resizing has been disabled.