googleanalytics extension: subdomain support, improved tracker configuration.
Added configuration members to indicate the domain for the Google Analytics
cookie (necessary on multiple domain / subdomain scenarios). Also, make it
possible to customize the parameters for the tracker instantiation when using
the universal API.

As a side effect: don't include the Analytics script / initialization blob if
there's no account defined.