Consider different languages when translating items in the backoffice.
Until now, the User.prefered_language member dictated the language used by both
UI elements and object descriptions in the backoffice. This caused a range of
problems, made more visible by the support for translation inheritance.

With this change, a chain of languages can be set up (on a per site or a per
user basis) that tells the backoffice which languages it should consider when
describing objects.