Replaced Publishable.translation_enabled with Publishable.enabled_translations.
Using a translated boolean to indicate the publication state for translations
had a fundamental problem: a translation had to *exist* in order to state wether
it was published. That used to be a perfectly fine assumption - if a piece of
content was not translated into a certain language, it was automatically not
published in that language. With the advent of translation inheritance, that no
longer holds true. An element may not directly define a translation for a
certain language, but it may still inherit a translation from a source locale.
The translation_enabled member, being a translatable member, made it difficult
to selectively enable / disable the publication of certain locales, as one would
have to add a translation for a derived locale just to disable it, which is
certainly counterintuitive.

The solution is to dettach publication state from translations, to a certain
degree. Instead of a translatable boolean, give all publishables a set of
locales that they should be published in (limited to the defined translations
plus their derived locales).

This was all made possible by having cocktail support collection indexing, which
was a prerequisite for fast resolution of IsPublishedExpression under the new
model.