Added the Item.get_representative_image() method.
This method makes it easier to have an item act as the representative image of
another item. Previous approaches included implementing a custom ImageRenderer,
but this was cumbersome and if done right it required defining event handlers to
listen to the lifecycle of the representative image, to invalidate the image
cache. It using a non direct reference, this was exponentially complex. Also, if
R is the representative image of object O, O.get_image_url() ==
R.get_image_url(), which wasn't true before (obviously, this has positive
effects on the cache and processing time).