Template.identifier shouldn't be a unique member.
Although normally it makes no sense to deliverately have two template objects
pointing to the same view, it can be necessary for historical reasons; also,
some sites or extensions can add members to the Template model (eg. the
newsletters extension adds slots for the header and footer, making it desirable
to have two or more templates with the same view and different headers/footers).