Show all possible user actions, toggle them as the selection changes.
Previously, listings only displayed the actions applicable to the selected
content type. In heterogeneous listings, this caused many actions to remain
hidden (ie. publishables in Item listings didn't offer a 'Show published'
action).

With the new system, the server tells the client beforehand which actions are
applicable to each entry in the listing, and the client toggles the button as
the selection changes.

A number of changes were necessary to achieve this:

- The add_view_action_context() and get_view_actions_context() functions have
been removed
- The behavior of ActionBar.action_context has changed; it is now a set of
strings, instead of a single string
- Added several methods to UserAction
- Contexts with an _extra suffix no longer work; use the new show_as_primary
or is_primary() members of UserAction instead
- Action buttons no longer receive a CSS class, they now get a data- attribute
instead