Use a compact "Element" column for most backoffice listings.
The descriptive "Element" added by the backoffice to all listings is mostly
redundant; it is only useful (or rather, indispensable) on heterogeneous
listings that don't offer a set of sufficiently descriptive members (basically,
Item and Publishable).

Existing sites used one of two solutions to deal with this:

- Hiding their descriptive members, using their listed_by_default property. The
  problem with this approach is that it the Element column is not sortable, nor
  does it show translations in different languages.
- Hiding the Element column, using the show_element_in_listings property. This
  works better... but hides the element's icon, and can lead to confusion (some
  listings have the column, some don't).

The new approach always adds the Element column, but reduces it to an icon /
thumbnail in most listings. Only listings for models with the new
requires_descriptive_handler property set to True will display the item's text.

As a side effect:
- Item.show_elements_in_listing is gone, replaced by the
  requires_descriptive_handler property
- ItemDisplay now inherits from ItemLabel, and its context menu functionality
  has been relocated to the new ItemContextMenu view
- ContentTable and ItemCollectionEditor rows now have context menus
- ItemCollectionEditor now uses backoffice_element_column_display as its UI
  generator