Removed the "Type" column from backoffice listings.
Even in heterogeneous listings, the name and properties of the elements, plus
the use of the contextual menu and/or the thumbnail column will most times be
enough to discern the type of each row.

Also, since it was a synthetic column, there was no way to specify its position
relative to the other columns, and therefore it was always displayed as the
first column after the thumbnail; this was inconvenient in many listings.