Refactored the woost.controllers.otifications API.
Instead of using tuples to represent notifications, use instances of the new
Notification class. The API remains mostly compatible with the previous one; the
only incompatible changes are in the NotificationBox view ("notification"
elements renamed to "notification_entry", and they now receive a Notification
object instead of a tuple).

The new API makes it possible to implement more complex notifications, such as
notifications with alternative views (for example, to offer embeded controls) or
with arbitrary state.