Fixed a regression in ItemCollectionEditor (keyboard navigation was broken).
The change to add contextual menus changed the way by which checkboxes were
hidden: relying on "overflow: hidden" was no longer an option. But using
"display: none" prevented the checkboxes from gaining focus, and that broke
the keyboard navigation of the control.

The adopted solution consists in adding a wrapping element for each checkbox with
"overflow: hidden", thus allowing other elements in the row (ie. the contextual
menu) to overflow.