Added the Menu.item_is_hidden() method.
Sometimes it is necessary to show or hide entries on a menu based on complex
filters that the "all or nothing" nature of Publishable.hidden member doesn't
cover. The new method makes it possible to override this logic, while still
preserving the visibility rules imposed by permissions and publication state.