locations extension: Added the CountryReference member type.
Many sites that use the locations extension do so to add references to
countries. The CountryReference member makes it easier to define these kinds
of references. As an added bonus, it provides a normalization rule so its value
can be set to an ISO code string (very useful when defining defaults).