Changed the point where the language component is added / removed from URLs.
If a content with per_language_publication == True is requested and the URL
contains no language identifier, the CMS adds the component and redirects the
client. Conversely, if per_language_publication == False, any language component
will be removed and the client will be redirected.

Until now, this check was performed during controller resolution. The problem
with that is that the cocktail dispatcher traps redirections and supresses them
momentarily, in order to trigger events in the whole request handler stack (f.
eg. before_request, exception_raised, etc). The event handler for the
before_request event in CMSController would proceed to validate the current
publishable, and an authorization error here would prevent the original
redirection from being raised once again, which prevented the language component
from being added / removed.

To fix this, the addition or removal of the language component is now done on
the before_request event handler of CMSController.