Refactored the objectio module to improve object identification across sites.
The previous implementation relied on the global_id and qname members alone;
that is fine in many cases, but fails to contemplate other situations:

- Users should be matched by email
- Countries (ie. from the locations extension) should be matched by their ISO
  code

And so on. This changeset modifies the import / export procedure to achieve two
things:

- As well as through qname and global_id, objects will now be matched using any
  of their unique members
- Classes can customize the matching procedure

Data exported with the previous method will not be compatible with importers
based on this changeset.

The following changes have been required to achieve this:

* Added generic methods to allow models to customize their reference generation
  and resolution procedure:
  - get_object_ref()
  - resolve_object_ref()
  - create_object_from_ref()
* Added Exporter._get_object_ref()
* Importer.resolve_ref() has been renamed to Imported.resolve_object_ref()
* The signature for Importer._handle_missing_object() has changed