googleanalytics: changed the way to define custom dimensions and metrics.
Instead of relying on the declaring_tracker event (which is still possible),
site implementors can use the new GoogleAnalyticsCustomDefinition model, which
provides a more obvious, centralized and declarative way of creating custom
dimensions and metrics.

Also, a future changeset will apply the values exported by this definition to
the target of Google Analytics events generated by the add_event() function.