googleanalytics: make it easier to override the values for custom definitions.
Added the GoogleAnalyticsCustomDefinition.identifier member and changed the
signature of utils.get_ga_custom_values(). The changes make it possible to
obtain a mapping of values for the site's custom definitions without supplying a
publishable object, and to override the value for specific definitions.

Moreover, value overrides can be specified using the definition's identifier,
instead of their usual numerical index, which makes the code much more readable
and resistant to all the shenanigans caused by Google Analytic's numerical
sequencing of definitions.