googleanalytics: declare events on links using a data- attribute.
Previously add_event() added a call to woost.ga.triggerEventOnClick() to its
target. The change relies on a cocktail binding an a data-woost-ga-event
attribute instead. This declarative approach is more elegant, reduces clutter on
the produced HTML and makes it easier to see event data at a glance from DOM
inspectors.