Check availability when restoring the language preference from a cookie.
The check makes sure that the language indicated by the cookie exists on the
site and is enabled. Before this change was applied, limiting enabled languages
by setting the 'published_languages' of Configuration or Website would cause
users of previously enabled languages to see a 404 page when returning back to
the site.