Reorganization of the Block model to support different views per block type.
Block types such as PublishableListing, NewsListing and others shadowed the
Block.view_class attribute with a string member, allowing editors to change the
template applied to blocks. The problem with this approach is that each block
type had to implement this on its own. This change provides a standard feature
(the Block.views attribute) to make this available to all block types, out of
the box.

This change has also brought about the following changes:

- Block.element_type is now a standard member of all blocks (most blocks ended
  up adding it anyway). The ElementType member type is gone.

- Block.heading_type remains, but some of its values now reside in a new
  Block.heading_display member

- Blocks with a visible heading now default to a section/h1 combination

- Most block members can be set to a "-- Default --" value, intended to respect
  the values set by the chosen view class

- Rearranged the fieldsets / tabs for Block, TextBlock and others. The
  "behavior" member group is now "publication".

- AddBlockNode is replaced by BlockEditNode