Don't assume that all requests will match a Website.
When websites mapped 1-1 to hostnames, it was reasonable to expect that all
requests would match a single website. Configurable URL mapping breaks this
assumption (since websites may be mapped to any other part of the URL - such as
the path - which may or may not be present - or even desirable).