Don't persist read only data in the edit stack.
When pickling the edit stack to store in the user's session, the form_data
attribute would be stored whole, including read only members. This is pointless
(since the values for read only members are injected into form_data on each
request anyway) and could lead to performance issues or exceptions for excessive
recursion.