Introduced default document templates.
Until now, each document had to specify its template explicitly. This presented
two problems:

* It was wasteful. In most cases, all documents of a type used the same
  template. Having each instance specify their template was inefficient
  (template objects ended up with a list containing all related documents using
  the template, which could escalate quickly).

* It made it difficult to choose templates based on dynamic or contextual
  parameters. For example, using different templates depending on the active
  website.

The new system still offers the Document.template member, which works as before,
but couples it with new elements that make it possible to specify site wide
default templates for all instances of a document type:

* Added the Document.get_template() method
* Added the Document.get_default_template() method
* Added settings in Configuration and Website to specify the default templates
  for all standard Document subclasses (default_page_template,
  default_news_template and default_event_template)

As a side effect, a new member subgroup has been added to both Configuration and
Website (presentation.appearence), and the 'theme' member for both has been
relocated inside it.

Implementers should replace snippets like this:

    default_template = schema.DynamicDefault(
        lambda: Template.get_instance(qname = "...")
    )

with an override of the get_default_template() method.

A migration is provided to automatically remove redundant references to a
default template. It does a best effort attempt; site developers should check
the results and adjust as needed.