Make publication state checks more flexible.
Until now, publication checks (the is_published() / is_accessible() methods of
the Publishable model, and the IsPublishedExpression / IsAccessibleExpression
classes) only considered the active language and website when computing their
results. This changeset makes it possible to indicate which languages and
websites should be considered. The default behavior will still use the active
context, so the changes should be backwards compatible (except for the "any"
value in the methods; see below).

* Make the 'language' and 'website' parameters to Publishable.is_published()
  and Publishable.is_accessible() accept multiple values, as well as the new
  Publishable.any_language / Publishable.any_website constants.
* Added the 'language' and 'website' properties and constructor parameters to
  IsPublishedExpression and IsAccessibleExpression
* The 'website' parameter of is_published() / is_accessible() no longer accepts
  an "any" value; pass Publishable.any_website instead.