Automatically generate the robots.txt file, relocate robots_should_index.
Until now, disabling indexing for specific content was only possible for
instances of the Document model (since the CMS used a <meta> HTML attribute to
achieve this, therefore requiring an HTML template). With this change, the CMS
will now automatically generate a robots.txt file, based on the
robots_should_index member. As a consequence, the member has been moved up to
the Publishable model.

Sites can customize the generated robots.txt file using any of the events
provided by RobotsController, or outright disable the automatic generation by
placing a custom robots.txt file in the document root of the web server.