Cache small screen navigation by default.
The menu for small screens can be expensive to render on sites with a big page
tree. This changeset enables caching for it, making all pages share the same
rendering, which can lead to a substantial performance improvement. Note that as
part of the change, the menu now selects the active menu entry client side.

Note that the site will need to configure a storage for the
cocktail.html.rendering_cache to benefit from this.