Added a mechanism to specify the default controller for a model.
This adds a @with_default_controller() decorator that mirrors the behavior of
@with_default_template(). The following changes have been made:

- Added woost.models.with_default_controller()
- Use the decorator to add settings to establish the default controller for all
standard Publishable models
- Added the publication.controllers member group for Configuration and Website
- PublishableController now does everything that DocumentController did.
DocumentController is kept as an alias of PublishableController, for backwards
compatibility.
- Added get_theme() and get_template() methods to PublishableController,
replacing the get_document_theme() method and page_template property of
DocumentController
- Moved the redirection_mode, redirection_target and redirection_method members
from Document to Publishable
- Declare an empty "template" attribute on PublishableObject
- Declare an empty "theme" attribute on PublishableObject
- Moved get_template() and get_default_template() from Document to PublishableObject
- Added get_controller() and get_default_controller() methods to PublishableObject

This change requires a migration. To take full advantage, of the change, project
maintainers should look for patterns like this in their models:

default_controller = schema.DynamicDefault(...)

and replace them with calls to woost.models.with_default_controller().