Take all the extensions out of the core.
The rationale behind the change is to greatly reduce the cost of applying
backwards incompatible changes to the core; extensions will only be updated to
cope with breaking changes as the demand arises.

Other benefits include having independent versioning for each extension, using
the Python package system.

- Removed the Extension model
- Added the ExtensionsManager class
- Extensions should now be implemented using an independently distributed
subpackage of the woost.extensions namespace package