Split User.prefered_language into the preferred_language and backoffice_language members
Beyond fixing the typo in the original name, the split makes it possible to
store a general language preference for each user, whereas before it only
applied to the backoffice, which frequently offers a much narrower breadth of
language choices than the frontend. Also, preferred_language now defaults to
the language that was active when the user was created, thus remembering their
language choice automatically.

Note that preferred_language is not used by Woost at all (though this might
change in the future), it is just recorded so that applications can use it were
necessary (for instance, to choose the language for an EmailTemplate).