Make the blocks preview refresh the whole slot if it can't find the element for a given block
New TextBlock instances weren't displayed on the preview view, because they
start with no text and therefore they are not rendered (the server sets them to
visible = False). Once the user gave them text, they still didn't show, because
the preview couldn't find the DOM element to update, as it wasn't present. Now
updateBlock messages pass slot information, so that the preview can fall back to
updating the whole slot.