Changes to the publication state algorithm
In order for an object to be published PublishableObject.enabled must now be
True. Previously, it was only considered if per_language_publication was False,
but this made disabling translated objects cumbersome, since editors had to
empty enabled_translations or set per_language_publication to False.

Beyond the UX benefits, the change should also help performance a bit, as it
removes some set intersection operations.