Support limiting the type of blocks accepted by slots
The general aim is to enforce certain blocks from being placed in an
inappropiate container, and limiting the blocks catalog offered to editors
depending on the context. The driving use case for this is newsletters, which
require specific blocks (ie. based on HTML tables) that are not suitable on
ordinary documents, and viceversa.

Changes:
- Added the Block.blockSubsets property
- Added the Slot.subset property
- Propagate both properties to the admin client side model descriptions
- Change the block palette depending on the context
- Prevent drag & drop of incompatible blocks