Changed the way per block embedded styles are assembled
Previously, embedded_styles_initialization was only considered if the block had
a value for embedded_styles. This is problematic for blocks that add content
that is positioned outside the block itself (f. eg. dynamic content added to the
<body> root by a script). After the change, blocks will now add an embedded
styles block if they define either of these members (so, in short, it is now
possible to have a block with embedded_styles_initialization but no
embedded_styles).

This is a forward port of changeset cd2bfd8466b9 in the quake branch (a graft
didn't make much sense, since the code for the style embedding has diverged from
the original)