Support registering admin views without specifying a target
This makes it possible to define views that should only be used by specific
nodes. Previously this required those nodes to be a section defined server side;
this is not always the case, as some nodes may be spawned dynamically client
side, and they may require a certain view that doesn't make sense anywhere else.