Added the View.extra_members property
This greatly simplifies the process of adding computed columns to a listing, but
it has required quite a bit of refactoring. Mainly, views are no longer exported
along admin data as JSON, but rather defined by SchemasController as Javascript
snippets (because they need to export member definitions to the client).

More changes:

- Added the ListingController.get_member()
- Added the Export.extra_members property
- Make schemaexport and dataexport assume read/write permissions on members
  without a schema
- Changed the resolution of the ui_component in the client to support callables