Support for managed roles and the General role
Added the Role.managed_roles member (and its related end Role.overseeing_roles)
that makes it easy to indicate that users of a role can only create / modify
/ delete users belonging to a subset of roles.

Although this was possible by using a set of carefully implemented permissions,
the new member makes it much simpler.

In order to implement this it was necessary to add a woost.general_role Role
that is given implicitly to all users as their first role (by
User.iter_roles()). This makes it possible to define permissions that are
applied before any other role is evaluated, thus setting site wide rules.